rt-app_parse_config: Change thread event handler to table-driven
The parse_thread_event_data function is becoming rather too long to
read easily. Lets separate out the individual event functions and
associate event->type with event name and event function in a table we
can loop through. This should be a bit easier to read.

Also included a call to check that the ordering of the table is valid
when rt-app starts.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>