uclamp: Add sched_setattr uclamp_hold support to rt-app like other attrs
uclamp is gaining a new hold (sometimes called minimum activation time)
attribute which will impose a minimum amount of time in which the apparent
utilization of a runqueue is guaranteed, triggered by the wakeup of a task
with the correct attribute.

Add support just like we have for the other per-task util-clamp attributes.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>